% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{apply.season}
\alias{apply.autumn}
\alias{apply.season}
\alias{apply.seasonally}
\alias{apply.spring}
\alias{apply.summer}
\alias{apply.winter}
\title{对xts时间序列按季节（如按春夏秋冬或者全部季节）应用函数。}
\usage{
apply.seasonally(x, FUN, ...)

apply.spring(x, FUN, ...)

apply.summer(x, FUN, ...)

apply.autumn(x, FUN, ...)

apply.winter(x, FUN, ...)
}
\arguments{
\item{x}{xts时间序列}

\item{FUN}{某R函数}
}
\value{
对该时间序列按季节应用FUN产生的结果
}
\description{
按3~5月为春季，6～8月为夏季，9~11月为秋季，12～次年2月为冬季的方法算。
}

