% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{BIAS}
\alias{BIAS}
\title{Relative bias}
\usage{
BIAS(sim, obs)
}
\arguments{
\item{sim}{Simulated data series or data to be evaluated.}

\item{obs}{Observed data series or data as reference.}
}
\value{
Relative bias of efficiency of sim and obs series.
}
\description{
Calculate relative bias (fraction, not percentage) of two data series.
}

