% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vic_relevents.R
\name{create_soil_params}
\alias{create_soil_params}
\title{Create the soil parameters for VIC model.}
\usage{
create_soil_params(coords, elev, soil_props, anprec, nlayer = 3, avg_T = NA,
  quarzs = NA, organic = NA, Javg = NA)
}
\arguments{
\item{coords}{A data frame containing ongitude and latitude coordinate of the gridcells.
First column must be the longitudes while the second column must be the
latitudes.}

\item{elev}{A vector containing the elevation of each gridcell.}

\item{soil_props}{A dataframe like what the function soil_convert needs.}

\item{anprec}{A vector containing the annual average precipitation of each gridcell.}

\item{nlayer}{Num of soil layers.}

\item{avg_T}{Average soil temperature of each gridcell. Would create from the global
dataset by interpolation if not offered.}

\item{quarzs}{Quarz content of soil of each gridcell. Would create from the global
dataset by interpolation if not offered.}

\item{organic}{Organic parameters. Including the fraction of organic matter, its bulk
density and its partial density for each soil layer.}

\item{Javg}{Average temperature in July at each gridcell.}
}
\value{
A data frame containing the soil parameters needed by the VIC model. Can
        be write to file use write.table for use.
}
\description{
Create the soil parameters for VIC model by offering coordinates, elevations
             and other datas. Each row of each data must be corresponding to a gridcell.
}
