% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vic_relevents.R
\name{create_veg_params}
\alias{create_veg_params}
\title{Create the veg params file for VIC model.}
\usage{
create_veg_params(pre_veg_params, out_file, aban = 0.03)
}
\arguments{
\item{pre_veg_params}{A vector of integer, the rare two bit means the id ofa kind of
vegetration conver, and the front bits means the id of the gridcell.
It's generally generated by Arcgis using the raster calculater.}

\item{out_file}{Path of the output veg_param file.}

\item{aban}{A fraction, when a kind of vegeraction cover fraction is little then it, it will
not be considerated in.}
}
\description{
Create the vegetration parameter file (generally named "veg_params.txt") with
             the output stat data of each single vegetration tile and the gridcell id which
             it distributed in. The vegetration data is usually from the dataset of UMD.
}

