% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO56.R
\name{delta_svp}
\alias{delta_svp}
\title{Estimate the slope of the saturation vapour pressure curve.}
\usage{
delta_svp(t)
}
\arguments{
\item{t}{Daily mean air temperature at 2m height [deg Celsius].}
}
\value{
Slope of the saturation vapour pressure curve.
}
\description{
Estimate the slope of the saturation vapour pressure curve
             by providing a air temperature.
}

