% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO56.R
\name{et0_hg}
\alias{et0_hg}
\title{Estimate ET0 by Hargreaves equation.}
\usage{
et0_hg(tmax, tmin, tmean = NULL, Ra = NULL, lat = NULL, dates = NULL)
}
\arguments{
\item{tmax}{Daily maximum air temperature at 2m height [deg Celsius].}

\item{tmin}{Daily minimum air temperature at 2m height [deg Celsius].}

\item{tmean}{Daily mean air temperature at 2m height [deg Celsius]. If not
provided it would estimated by averaging the tmax and tmin.}

\item{Ra}{Clear sky incoming shortwave radiation, i. e. extraterrestrial
radiation multiply by clear sky transmissivity (i. e. a + b,
a and b are coefficients of Angstrom formula. Normally 0.75)
[MJ m-2 day-1]. If not provided, must provide lat and dates.}

\item{lat}{Latitude [degree].}

\item{dates}{A R Date type of a vector of Date type. If not provided, it will
Regard the ssd series is begin on the first day of a year.}
}
\value{
Reference evapotranspiration ET0 from a hypothetical grass
        reference surface [mm day-1].
}
\description{
Estimate reference evapotranspiration (ET0) from a hypothetical
             short grass reference surface using the the Hargreaves equation.
}

