% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO56.R
\name{et0_pm}
\alias{et0_pm}
\title{Estimate ET0 by FAO-56 Penman-Monteithe quation.}
\usage{
et0_pm(Rs, tmax, tmin, ws, Rso, G = 0, h.ws = 10, albedo = 0.23,
  delta = NULL, gamma = NULL, z = NULL, ea = NULL, es = NULL,
  rhmean = NULL, pres = NULL)
}
\arguments{
\item{Rs}{Incoming shortwave radiation at crop surface [MJ m-2 day-1].}

\item{tmax}{Daily maximum air temperature at 2m height [deg Celsius].}

\item{tmin}{Daily minimum air temperature at 2m height [deg Celsius].}

\item{ws}{Wind speed [m s-1].}

\item{Rso}{Clear sky incoming shortwave radiation, i. e. extraterrestrial
radiation multiply by clear sky transmissivity (i. e. a + b,
a and b are coefficients of Angstrom formula. Normally 0.75)
[MJ m-2 day-1].}

\item{G}{Soil heat flux [MJ m-2 day-1]. Normally set to 0.0 when the time
steps are less than 10 days. Should be calculated in monthly or
longer time step. Default 0.0.}

\item{h.ws}{Height where measure the wind speed [m]. Default 10m.}

\item{albedo}{Albedo of the crop as the proportion of gross incoming solar
radiation that is reflected by the surface. Default 0.23
(i. e. the value of the short grass reference crop defined
by the FAO).}

\item{delta}{Slope of the saturation vapour pressure curve. Can be estimated
by air temperature if is not provided.}

\item{gamma}{Psychrometric constant [kPa Celsius -1]. Can be estimated
by the air pressure or elevation if is not provided.}

\item{z}{Elevation above sea level [m]. Must provided if gamma and pres
are not provided.}

\item{ea}{Actual vapor pressure [kPa]. Can be estimated by maximum or minimum
air temperature and mean relative humidity.}

\item{es}{Saturated vapor pressure [kPa]. Can be estimated by maximum or
minimum air temperature if is not provided.}

\item{rhmean}{Daily mean relative humidity [%]. If not provided then
the actual vapor pressure would be estimated by minimum air
temperature.}

\item{pres}{Air pressure [kPa]. If not provided, must provide z.}
}
\value{
Reference evapotranspiration ET0 from a hypothetical grass
        reference surface [mm day-1].
}
\description{
Estimate reference evapotranspiration (ET0) from a hypothetical
             short grass reference surface using the FAO-56 Penman-Monteithe
             quation (equation 6 in Allen et al. (1998))
}

