% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO56.R
\name{ext_rad}
\alias{ext_rad}
\title{Estimate daily daily extraterrestrial radiation.}
\usage{
ext_rad(lat, dates)
}
\arguments{
\item{lat}{Latitude [degree].}

\item{dates}{A R Date type of a vector of Date type. If not provided, it will
Regard the ssd series is begin on the first day of a year.}
}
\value{
extraterrestrial radiation [MJ m-2 day-1].
}
\description{
Estimate daily daily extraterrestrial radiation [MJ m-2 day-1].
}
