% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{FAR}
\alias{FAR}
\title{False alarm ratio}
\usage{
FAR(sim, obs, th = 0)
}
\arguments{
\item{sim}{Series to be evaluated.}

\item{obs}{Reference series.}

\item{threshold}{Threshold to determind the event if happened. Default 0.}
}
\value{
FAR of the evaluated series and the reference series.
}
\description{
False alarm ratio of the estimation, like precipitation.
}

