% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{find_date}
\alias{find_date}
\title{Get the subscript of the dates by providing the start date of the data.}
\usage{
find_date(orist, st, ed = st)
}
\arguments{
\item{orist}{start date of the data. Can be string or string-like integer.}

\item{st}{start date of the subset or a series of dates.}

\item{ed}{End date of the subset.}
}
\value{
A series of subscripts can be used to get the subset of data.
}
\description{
Get the subscript of the dates by providing the start date
             of the data so as to get the subset of the data at the certain
             dates.
}

