% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{fitPearson3}
\alias{fitPearson3}
\title{Fit the Pearson type III distribuion}
\usage{
fitPearson3(sfs, init = NULL)
}
\arguments{
\item{sfs}{Sequence of streamflow.}

\item{init}{Vector of initiate parameters of P-III distribution,
must be c(xm, Cv, Cs). It can estimate when not offering.}
}
\value{
A list including the estimated parameters, and a function
        to calculate the frequency of a certern flood (freq), and a
        function to calculate streamflow of a certern frequency (qnt).
}
\description{
Using a Pearson III distribution to fit the streamflow
             sequence.
}

