% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{geo2table}
\alias{geo2table}
\title{Make 2d grid-like data to station-like table data}
\usage{
geo2table(arr, dimxy = NULL, x = NULL, y = NULL, mask = NULL,
  xy = NULL)
}
\arguments{
\item{arr}{The array of the data to be transform.}

\item{dimxy}{Which dimensions are the x and y. Default c(1, 2), which
means the first and second dimensions are x and y.}

\item{x}{Coordinates of the x dimension.}

\item{y}{Coordinates of the y dimension.}

\item{mask}{An matrix with the same dim of x and y, marking which the
grids to be transform (NA means not be transform).}

\item{xy}{A series of coordinates of the grid cells to be transform.}
}
\value{
An list, including the array tranformed, and the corresponding
        coordinates.
}
\description{
Make the grid-like data to table-like data. For example,
             an nc data has four dimensions, longitude, latitude, height
             and time. Then it transfrom the data into a three dimensions
             data, with grid, height and time as dimensions.
             If the original data is a three dimension data (x, y and
             time) then it would turn it to a table, each row is a time
             step while each column is a grid.
}

