% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{get_nclon}
\alias{get_nclon}
\alias{get_nclat}
\alias{get_nctime}
\title{Get dimension values by one line code.}
\usage{
get_nclon(file, var = NULL)

get_nclat(file, var = NULL)

get_nctime(file, var = NULL, conv = TRUE)
}
\arguments{
\item{file}{File path of the netCDF file.}

\item{var}{The name of the variable where to get the dimension.}

\item{conv}{Convent the time values from number to time or not. Default TRUE.
Specified for function get_nctime.}
}
\value{
The values of the dimension of the netCDF file.
}
\description{
Get typical dimension values, like lon, lat and time by one line code.
}
