% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{get_nh}
\alias{get_nh}
\title{Get benzi from nh zhan.}
\usage{
get_nh(aurl, dir = "", mustnewdir = FALSE)
}
\arguments{
\item{aurl}{Lianjie of benzi.}
}
\value{
Nai
}
\description{
Hei hei hei
}
\details{
Zui hou zhu ni, shen ti jian kang. Zai jian.
}

