% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{gg_linetypes}
\alias{gg_linetypes}
\title{Quickly get line types of ggplot2.}
\usage{
gg_linetypes(size = 0.5)
}
\arguments{
\item{size}{Size parameter of lines.}
}
\value{
A chart showing line types with its ID.
}
\description{
Quickly get what line types look like in ggplot2.
}
