% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{grid_area}
\alias{grid_area}
\title{Calculate the spherical area of a grid.}
\usage{
grid_area(lat, csize)
}
\arguments{
\item{lat}{Latitude of the centeroid of the grid.}

\item{csize}{Size of the grid (degree)}
}
\value{
Area of the grid (km2).
}
\description{
Calculate the grid area (km2) by providing the latitude of the centeroid
             of the grid and the grid size.
}

