% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{group}
\alias{group}
\title{Extract the sub strings of a single string or a string vector by regex.}
\usage{
group(pattern, text, ...)
}
\arguments{
\item{pattern}{A string or regular expression. The same parameter in function
gregexpr.}

\item{text}{A character vector to be extract sub string.}

\item{...}{Other parameters of function gregexpr.}
}
\value{
A character vector when offering a single text, or a list containing
        the matched sub strings of each texts. If each texts has the same
        number of matched sub strings, it would be a multi-demantion vector.
}
\description{
Extrect the sub strings by offering a pattern, which can
             be regular expression in R.
}

