% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{list_ncvar}
\alias{list_ncvar}
\title{Show the variables or their dimensions of a netCDF file.}
\usage{
list_ncvar(file, show.dim = FALSE)
}
\arguments{
\item{file}{Path of the nc file.}

\item{show.dim}{If is TRUE, also list out the dimensions of each variables.
Default FALSE.}
}
\value{
A vector of the variable names.
}
\description{
Show the variables or their dimensions of a netCDF file.
}

