% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO56.R
\name{lw_rad}
\alias{lw_rad}
\title{Estimate Incoming longwave radiation.}
\usage{
lw_rad(temp, ea = NULL, s = 1, method = "MAR")
}
\arguments{
\item{temp}{Near surface air temperature [deg Celsius].}

\item{ea}{Near surface actual vapour pressure [kPa].}

\item{s}{Cloud air emissivity, the ratio between acutal incomming shortwave
radiation and clear sky incoming shortwave radiation.}

\item{method}{The method to estimate the air emissivity.
Must be one of 'MAR', 'SWI', 'IJ', 'BRU', 'SAT', and 'KON'.}
}
\value{
Incomming longwave radiation [W /m2].
}
\description{
Estimate Incoming longwave radiation. Not included in FAO56
             paper but added for convinience.
}

