% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{MK_test}
\alias{MK_test}
\title{Mann-Kendall trend test}
\usage{
MK_test(x, sen.slope = TRUE, p.value = TRUE)
}
\arguments{
\item{x}{A time series (vector or matrix) to be test.}

\item{sen.slope}{Output Sen's slope or not. Default TRUE.}

\item{p.value}{Output significance confidence value or not. Default TRUE.}
}
\value{
A named vector include M-K static value(Z), Sen's slope value
        (sen.slope) and significance confidence value (p) when x is
        a vector, or a matrix to record Z, sen.slope and p for
        each column of x.
}
\description{
Mann-Kendall trend test
}
