% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{nc_combine_time}
\alias{nc_combine_time}
\title{Combine several netCDF file by time.}
\usage{
nc_combine_time(fs, out_file, from = NULL, to = NULL, datum_nc = 1,
  compression = NA, slip = 30)
}
\arguments{
\item{fs}{A vector of string. File paths of each nc file to be combine.}

\item{out_file}{Output path of the new combined nc file.}

\item{from}{A vector indicating the start of the time range to be combine of each nc file.}

\item{to}{A vector indicating the end of the time range to be combine of each nc file.}

\item{datum_nc}{Which nc file to be regard as the datum, i.e. the attributes, dimensions
and variable names would inherit from it.}

\item{compression}{The same parameters of function ncvar_def. Define the compression level
of the nc file.}

\item{slip}{Provided if the time datum ("XXX" of the "days since XXX") of each file are not
equal. It appoints the times interval (in time step) between each nc file to be
combine. Default 30.}
}
\description{
Combine several netCDF file by time. Each netCDF file should have the
             same dimensions and variables.
}

