% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{nc_intercept}
\alias{nc_intercept}
\title{Clip a part from a netCDF file to another netCDF file.}
\usage{
nc_intercept(file, out_file, from = NULL, to = NULL, dims = NULL,
  vars = NULL, byval = TRUE, compression = NA)
}
\arguments{
\item{file}{A netCDF file to be clip.}

\item{out_file}{The netCDF file to store the cliped part of the original netCDF file.}

\item{from}{A vector of the border that where to begin to clip the variables at each
dimensions. The length should be as same as dims. If byval is TRUE, or
time should be the "date-like" integer in the format of YYYYMMDD, such as
that 19600101 means "1960-01-01".}

\item{to}{A vector of the border that where to stop to clip the variables at each
dimensions. The length should be as same as dims, details see from.}

\item{vars}{Variables to be clip. Default all the variables of the origin nc file.}

\item{byval}{Determine the from and to are values of the dimensions (e.g. longitude
values) or the subscript of the dimensions (e.g. the N th value of the
variable). Default TRUE, use the values of the dimensions.}

\item{compression}{The same agrument of function ncvar_def.}

\item{dim}{Names of the dimensions to be clip. Default are longitude, latitude and time.}
}
\description{
Clip a part from a netCDF file to another netCDF file by providing dimensions, from an to.
}

