% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{ncdim}
\alias{ncdim}
\alias{ncdim_lat}
\alias{ncdim_lon}
\alias{ncdim_time}
\title{Quick build netCDF typical dimensions.}
\usage{
ncdim_lon(vals = NULL, from = NULL, csize = NULL, len = NULL,
  name = "lon", longname = "longitute")

ncdim_lat(vals = NULL, from = NULL, csize = NULL, len = NULL,
  name = "lat", longname = "latitute")

ncdim_time(time, name = "time", calendar = "proleptic_gregorian", ...)
}
\arguments{
\item{vals}{Values of the dimension. For lon and lat can be a series of longitudes and
latitudes, for time can be a series of dates and times. If not provided,
can provide from, csize and len for substitudes.}

\item{from}{Beginning of the values. Provided if vals was not provided.}

\item{csize}{Size of the gridcells for lon and lat. Also can be regarded as the interval
of the lon and lat values.}

\item{len}{Length of the values.}

\item{name}{Name of the dimensions. Default are lon, lat and time.}

\item{longname}{Long name of the dimensions.}

\item{time}{Time or date series for time dimension.}

\item{...}{Other parameter for function time2num.}
}
\value{
A netCDF dimension object.
}
\description{
Quick build typical netCDF dimensions like longitude, latitude and time.
}

