% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{NMSE}
\alias{NMSE}
\title{Normailzed mean square error}
\usage{
NMSE(sim, obs)
}
\arguments{
\item{sim}{Simulated data series or data to be evaluated.}

\item{obs}{Observed data series or data as reference.}
}
\value{
Normailzed mean square error of sim and obs series.
}
\description{
Calculate normailzed mean square error of two data series.
}
