% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{NSCE}
\alias{NSCE}
\title{Nash-Sutcliffe coefficient of efficiency}
\usage{
NSCE(sim, obs, ...)
}
\arguments{
\item{sim}{Simulated data series or data to be evaluated.}

\item{obs}{Observed data series or data as reference.}
}
\value{
Nash-Sutcliffe coefficient of efficiency of sim and obs series.
}
\description{
Calculate Nash-Sutcliffe coefficient of efficiency of two data series.
}

