% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO56.R
\name{ol_rad}
\alias{ol_rad}
\title{Estimate net outgoing longwave radiation.}
\usage{
ol_rad(tmax, tmin, Rs, Rso, ea)
}
\arguments{
\item{tmax}{Daily maximum air temperature at 2m height [deg Celsius].}

\item{tmin}{Daily minimum air temperature at 2m height [deg Celsius].}

\item{Rs}{Incoming shortwave radiation at crop surface [MJ m-2 day-1].}

\item{Rso}{Clear sky incoming shortwave radiation, i. e. extraterrestrial
radiation multiply by clear sky transmissivity (i. e. a + b,
a and b are coefficients of Angstrom formula. Normally 0.75)
[MJ m-2 day-1].}

\item{ea}{Actual vapor pressure [kPa]. Can be estimated by maximum or minimum
air temperature and mean relative humidity.}
}
\value{
Net outgoing longwave radiation [MJ m-2 day-1]
}
\description{
Estimate net outgoing longwave radiation.
}

