% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{p_linetypes}
\alias{p_linetypes}
\title{Quickly get line types of base plot.}
\usage{
p_linetypes()
}
\value{
A chart showing line types with its ID.
}
\description{
Quickly get what the line types look like in
base plot function.
}
