% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{p_pointshapes}
\alias{p_pointshapes}
\title{Quickly get point shapes of base plot.}
\usage{
p_pointshapes()
}
\value{
A chart showing point shapes with its ID.
}
\description{
Quickly get what shapes of the point shapes look like in
base plot function.
}
