% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{POD}
\alias{POD}
\title{Probability of the detection}
\usage{
POD(sim, obs, th = 0)
}
\arguments{
\item{sim}{Series to be evaluated.}

\item{obs}{Reference series.}

\item{threshold}{Threshold to determind the event if happened. Default 0.}
}
\value{
POD of the evaluated series and the reference series.
}
\description{
Probability of the detection, like precipitation.
}
