% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{point_shapes}
\alias{point_shapes}
\title{Quickly get point shapes.}
\usage{
point_shapes(size = 7)
}
\value{
A chart showing point shapes with its ID.
}
\description{
Quickly get what shapes of the point shapes look like in
ggplot2.
}

