% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{qplot_grid}
\alias{qplot_grid}
\title{qplot of the grid data.}
\usage{
qplot_grid(x, y, z, ..., data = NULL, facets = NULL)
}
\arguments{
\item{x, y}{Coordinates of the grid cells.}

\item{z}{Values of the grid cells.}

\item{...}{Aesthetics passed into each layer.}

\item{data}{Data frame to use (optional).}

\item{facets}{Faceting formula to use.}
}
\description{
A qplot function of ggplot2 for grid data.
}

