% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{radiation}
\alias{radiation}
\title{Calculate shortwave radiation using sunshine duration.}
\usage{
radiation(ssd, lat, ts = NULL, a = 0.25, b = 0.5)
}
\arguments{
\item{ssd}{Sunshine duration (hour) in a day.}

\item{lat}{Latitude of the station location.}

\item{ts}{Date or date series of the ssd data. Would regard the first
day in a year as the first ssd data if not provided.}

\item{a}{Coefficient of the Angstrom formula. Determine the relationship
between ssd and radiation. Default 0.25.}

\item{b}{Coefficient of the Angstrom formula as a.}
}
\description{
Calculate shortwave radiation using sunshine duration using
             Angstrom formula at daily scale.
}
\references{
Martinezlozano J A, Tena F, Onrubia J E, et al. The historical
            evolution of the Angstrom formula and its modifications: review
            and bibliography.[J]. Agricultural & Forest Meteorology, 1985,
            33(2):109-128.
}

