% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{read_ascgrid}
\alias{read_ascgrid}
\title{Read ArcInfo ASCII grid data.}
\usage{
read_ascgrid(file)
}
\arguments{
\item{grid_file}{A ArcInfo ASCII grid file.}
}
\value{
A grid data in list type, including x, y, xllcorner,
        yllcorner and cellsize.
}
\description{
Read the grid data from ArcInfo type ASCII file.
}
