% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{read_ncvar}
\alias{read_ncvar}
\title{Read data from nc file by one line code.}
\usage{
read_ncvar(file, var = NULL, from = NULL, to = NULL, dims = NULL,
  lon = NULL, lat = NULL, time = NULL, byval = TRUE, msize = 2048)
}
\arguments{
\item{file}{String. File path of the netCDF file.}

\item{var}{String. Name of the variable to be get data.}

\item{from}{A vector indicating where to start reading the values at each dimentions.
If byval = TRUE, it's regarded as the dimention value; If byval = FALSE,
it's regarded as the indexes.}

\item{to}{A vector indicating the end of the values to be reading at each dimentions.
If byval = TRUE, it's regarded as the dimention value; If byval = FALSE,
it's regarded as the indexes.}

\item{dims}{A vector indicating subset at which dimensions.}

\item{lon}{A length 2 vector. Indicating the longitude range of the data to be get.}

\item{lat}{A length 2 vector. Indicating the latitude range of the data to be get.}

\item{time}{A length 2 vector. Indicating the time range of the data to be get.}

\item{byval}{If set to TRUE, values of from and to would be regarded as dimension
values; If False, it would be regarded as indexes. Default TRUE.}

\item{msize}{Maximum size (MB) of the data to be read. This is provide that too much data
read into memory. If < 0 it will regarded as unlimited. Default 2048 MB (2GB).}
}
\value{
The values of the data of the netCDF file.
}
\description{
Using one line code to get data from netCDF file.
}

