% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{readIMERG_nc4}
\alias{readIMERG_nc4}
\title{Read IMERG data from netCDF4 file.}
\usage{
readIMERG_nc4(files, x, y, varsn = 1, verbose = FALSE)
}
\arguments{
\item{files}{Paths of the IMERG files to be read.}

\item{x}{Longitudes of the points where get the precipitation data.}

\item{y}{Latitudes of the points where get the precipitation data. Length
should be as same as x.}

\item{varsn}{Order of which variable to be read. Default the first variable.}

\item{verbose}{If TRUE, it will print the files read. Default FALSE.}
}
\value{
A data frame of the IMERG data, for each column means a point, and
        each row means a time step. If only offered one point or one file,
        will return a vector of the data.
}
\description{
Only for the nc4 format IMERG data. Offering a series of path
             of the IMERG files, or single file path, and the coordinates
             of the data.
}
