% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{show_pointshapes}
\alias{show_pointshapes}
\title{Quickly get point shapes.}
\usage{
show_pointshapes(size = 7)
}
\arguments{
\item{Size}{parameter of points.}
}
\value{
A chart showing point shapes with its ID.
}
\description{
Quickly get what shapes of the point shapes look like in
ggplot2.
}

