% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vic_relevents.R
\name{soil_convert}
\alias{soil_convert}
\title{Create the hydraulic parameters of soil from soil property parameters for VIC model.}
\usage{
soil_convert(soil_props, nlayer = 3)
}
\arguments{
\item{soil_props}{A dataframe containing the soil hydraulic parameters, including
volume fraction of gravity (fraction, not percentage), weight fraction
of sand, silt and clay, reference bulk density, and weight fraction of
organic matter, total 6 property parameters. The first 6 columns are
parameters of the first soil layer, and the next 6 columns are of the
second soil layer, and so on.}

\item{nlayer}{Num of layers of each gridcell.}
}
\value{
A dataframe containing the soil hydraulic parameters for each layer of each
        gridcell. Including EXPT (Exponent), Ksat (Saturated hydrologic conductivity, mm/d),
        BUBLE (Bubbling pressure, cm), BULKDN (Bulk density), WcrFT (Fractional soil
        moisture content at the critical point) and WpFT (Fractional soil moisture
        content at the wilting point).
}
\description{
Create the hydraulic parameters of soil from soil property parameters for VIC model.
}
\details{
This function is using the Saxton's formulars. If the soil property parametersis
         HWSD, The property parameters are corresponding to T_GRAVEL (fraction), T_SAND
         (fraction), T_SILT (fraction), T_CLAY (fraction), T_REF_BULK_DENSITY (g/cm3),
         and T_OC (fraction).
}
\references{
Saxton, K. E, Rawls, W. J. Soil Water Characteristic Estimates by Texture
            and Organic Matter for Hydrologic Solutions[J]. Soil Science Society of
            America Journal, 2006, 70(5):1569-1578.
}

