% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO56.R
\name{soil_heat_flux}
\alias{soil_heat_flux}
\title{Estimate monthly soil heat flux.}
\usage{
soil_heat_flux(t.p, t.n = NULL, t.c = NULL)
}
\arguments{
\item{t.p}{Mean air temperature of the previous month [Celsius].}

\item{t.n}{Mean air temperature of the next month [Celsius].}

\item{t.c}{Mean air temperature of the current month [Celsius].}
}
\value{
Soil heat flux [MJ m-2 day-1].
}
\description{
Estimate monthly soil heat flux (G) from the mean air
             temperature of the previous, current or next month assuming
             as grass crop.
}
