% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{statistic_col}
\alias{statistic_col}
\alias{sum_col}
\alias{mean_col}
\alias{max_col}
\alias{min_col}
\alias{var_col}
\alias{sd_col}
\title{Statistics of each respective column}
\usage{
sum_col(x, ...)

mean_col(x, ...)

max_col(x, ...)

min_col(x, ...)

var_col(x, ...)

sd_col(x, ...)
}
\arguments{
\item{x}{a matrix of dataframe}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
A series of statistic value of each column of x.
}
\description{
Returns the statistic values, such as sum, mean, max,
min of each column of dataframe or matrix.
}
