% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO56.R
\name{sur_rad}
\alias{sur_rad}
\title{Estimate daily solar radiation.}
\usage{
sur_rad(ssd, lat, dates = NULL, a = 0.25, b = 0.5)
}
\arguments{
\item{ssd}{sunshine duration [hours].}

\item{lat}{Latitude [degree].}

\item{dates}{A R Date type of a vector of Date type. If not provided, it will
Regard the ssd series is begin on the first day of a year.}

\item{a}{Coefficient of the Angstrom formula. Determine the relationship
between ssd and radiation. Default 0.25.}

\item{b}{Coefficient of the Angstrom formula. Default 0.25}
}
\value{
Solar radiation at crop surface [MJ m-2 day-1].
}
\description{
Estimate daily solar radiation at crop surface [MJ m-2 day-1]
             by providing sunshine duration (SSD) in hours.
}
\references{
Martinezlozano J A, Tena F, Onrubia J E, et al. The historical
            evolution of the Angstrom formula and its modifications: review
            and bibliography.[J]. Agricultural & Forest Meteorology, 1985,
            33(2):109-128.
}

