% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{table2geo}
\alias{table2geo}
\title{Make table-like array to 2d geo-like array}
\usage{
table2geo(arr, x, y, csize = NULL, dimpnt = NULL, dgt = 5)
}
\arguments{
\item{arr}{The array of the data to be transform.}

\item{x, y}{Coordinates of the grids.}

\item{y}{Coordinates of the y dimension.}

\item{csize}{Cell size of the grids.}

\item{dimpnt}{Site of the dimension difine the location of grids.}
}
\value{
An list, including the array tranformed, and the corresponding
        coordinates.
}
\description{
Make the table-like data to grid-like data. For example,
             an arrray has three dimensions, station, height and time.
             Then it transfrom the data into a four dimensions
             data, with longitude, latitude, height and time as
             dimensions. If the original data is a 2d data (grid and
             time) then it would turn it to a 3d array. the coordinates
             of the grids must be grid-like.
}

