% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO56.R
\name{vp.temp}
\alias{vp.temp}
\title{Estimate saturation vapor pressure.}
\usage{
vp.temp(t)
}
\arguments{
\item{t}{Air temperature [Celsius].}
}
\value{
saturation vapor pressure [kPa].
}
\description{
Estimate saturation vapor pressure from air
             temperature.
}

