% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{write.meta4}
\alias{write.meta4}
\title{Save urls to be download as meta4 format.}
\usage{
write.meta4(urls, file = "")
}
\arguments{
\item{urls}{Url list of the file to be downloaded, such as OpenDAP.}

\item{file}{either a character string naming a file or a connection open
for writing. "" indicates output to the console.}
}
\description{
Save the url list of a series of file to be downloaded as
             .meta4 format for downthemall. No one can doubt that downthemall
             is a good add-on to make download more eaiser, convinient and
             managable. However, there are also some shortcommings to
             make it shit, for example, it would automatically remove the contents
             after square brackets without any consults when batch import
             url list, but those contents are necessary for some file urls such as
             OpenDAP. This function is to save the urls as meta4 file so
             that downthemall could not amend the urls when import them to
             the queue.
}

