% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{write_arc_grid}
\alias{write_arc_grid}
\title{Write ArcInfo ASCII grid data.}
\usage{
write_arc_grid(grid, out_file, xcor = NULL, ycor = NULL, csize = NULL,
  NA_value = -9999)
}
\arguments{
\item{grid}{A list of grid data or a matrix. When it's a matrix, xcor, ycor and
csize are necessary.}

\item{out_file}{Output file path.}

\item{xcor}{X corner of the grid. Necessary when grid is a matrix.}

\item{ycor}{Y corner of the grid. Necessary when grid is a matrix.}

\item{csize}{Cell size of the grid. Necessary when grid is a matrix.}

\item{NA_value}{A value to reprent the NA value.}
}
\description{
Write the grid data to the file as a ArcInfo type ASCII file by
             offering a list of grid data like the output of points2grid()
             or create_flow_direction(), or a matrix and some necessary
             informations like cell size, x and y corner of the grid data and
             so on.
}
