% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{write_ncvar}
\alias{write_ncvar}
\title{Write data to netCDF file.}
\usage{
write_ncvar(val, file, var, dims = c(), newvar = FALSE, units = "NULL")
}
\arguments{
\item{val}{Values to be write to the file. Should be matrix or array.}

\item{file}{Path of the netCDF file to be write.}

\item{var}{Name of the variable to be write. If the variable is not exist and
the parameter "newvar" set to TRUE then the variable would be create.}

\item{dims}{A vector of string to determine the dimensions of the variable
when create a new variable. If some dimensions was not determine,
it would atuomatically find a dimension that fit it.}

\item{newvar}{If is TRUE, if the variable is not exist, it would create a
new variable. Default FALSE.}

\item{units}{Units of the new variable if the new variable is created.}
}
\description{
Write data to netCDF file.
}

