% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{wt}
\alias{wt}
\title{write.table without col names and row names}
\usage{
wt(x, file = "", ...)
}
\description{
The original write.table set col.names and row.names TRUE,
and when we want to merely write out a pure dataset or matrix we should
write in a long line: "col.names = F, row.names = F" and it so inconvenient
and troublesome. Now have this function we could not write it again.
Other usage is as same as write.table.
}

