% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{apply.season}
\alias{apply.season}
\alias{apply.seasonally}
\alias{apply.spring}
\alias{apply.summer}
\alias{apply.autumn}
\alias{apply.winter}
\title{Apply function over seasons for xts object.}
\usage{
apply.seasonally(x, FUN, ...)

apply.spring(x, FUN, ...)

apply.summer(x, FUN, ...)

apply.autumn(x, FUN, ...)

apply.winter(x, FUN, ...)
}
\arguments{
\item{x}{An xts time-series object}

\item{FUN}{An R function}

\item{...}{Additional arguments to FUN.}
}
\value{
The result of the R function apply to the xts object.
}
\description{
The seasons is conclute as that March, April and May as spring,
             June, July and August as summer, September, October and November
             as autumn, December and January, Feburary of next year as winter.
}
