% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{loss_time}
\alias{loss_time}
\title{Get the deficiency of the time series.}
\usage{
loss_time(ts)
}
\arguments{
\item{ts}{Time series to be find the lose time.}
}
\value{
The time stamp lost in the time series.
}
\description{
Get the deficiency of the time series.
}
