% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{MK_mut_test}
\alias{MK_mut_test}
\title{Mann-Kendall mutation test}
\usage{
MK_mut_test(x, plot = TRUE, out.value = FALSE, index = NULL, p.size = 3,
  l.size = 0.6)
}
\arguments{
\item{x}{A single time series (can be xts object) to do that test.
If x has multi column, it will only choose the first column to do that test.}

\item{plot}{TRUE or FALSE. If it will draw a plot.}

\item{out.value}{TRUE or FALSE. If it return the UF and UB value series or plot.}

\item{index}{Index for x axis of plot.}

\item{p.size}{Point  size of plot.}

\item{l.size}{Line size of plot.}
}
\value{
Plot or values of UF and UB series from M-K mutation test.
}
\description{
Make Mann-Kendall mutation test and draw plot. Where the UF and UB curve cross means a mutation may be happen.
}
