% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{MK_test}
\alias{MK_test}
\title{Mann-Kendall趋势分析}
\usage{
MK_test(x)
}
\arguments{
\item{x}{待分析的时间序列，可以是数据框或者是矩阵形式}
}
\value{
如果x为单个序列就返回一个列表，其中MK.value为M-K统计值，p.value为显著性达到了多少置信度。
        如果x是个数据框或者矩阵就挨列计算，返回一个矩阵记录每一列的MK.value和p.value。
}
\description{
Mann-Kendall趋势分析
}

