% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibadaR.R
\name{MK_test}
\alias{MK_test}
\title{Mann-Kendall trend test}
\usage{
MK_test(x)
}
\arguments{
\item{x}{A time series (vector or matrix) to be test.}
}
\value{
A list include M-K static value(MK.value) and significance value(p.value)
        if x is a vector, or a matrix to record MK.value and p.value for each column of x.
}
\description{
Mann-Kendall trend test
}

