% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4_relevents.R
\name{nctime}
\alias{nctime}
\alias{num2time}
\alias{time2num}
\title{Convert numbers to time or convert time to numbers in nc format.}
\usage{
time2num(time, since = time[1], freq = NULL, tz = "GMT")

num2time(val, units = NULL)
}
\arguments{
\item{time}{R time or date series.}

\item{since}{Time datum of the time series after converted to numbers. Default the first
step of the time series.}

\item{freq}{Time step of the time series. Could select a value when not provided.}

\item{tz}{Time zone of the time series. Default UTC.}

\item{units}{NetCDF time units for function num2time.}
}
\value{
A list including the numeric time series and the netCDF style time unit for function
        time2num, or an R time series for function num2time.
}
\description{
Convert numbers of netCDF time dimension style to R time series by providing
             units or Convert R time series to numbers and units of netCDF time dimension
             style.
}

